requirements.txt
aiogram==3.27.0
python-dotenv==1.0.1
Pillow==10.4.0
